pyproject.toml
[tool.poetry]
name = "catalogues"
version = "0.1.0"
description = "Сервис справочников КНД"
authors = ["rsaleev <rs.aleev@gmail.com>"]
packages = [
    {include = "./src"}
]

[tool.poetry.dependencies]
python = "^3.9"
uvloop = "^0.16.0"
fastapi = "^0.73.0"
fastapi-health = "^0.4.0"
tortoise-orm = "^0.18.1"
fastapi-utils = "^0.2.1"
asyncpg = "^0.25.0"
cachetools = "^5.0.0"
fastapi-cache2 = "^0.1.8"

[tool.poetry.dev-dependencies]
uvicorn = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
